handle fatal kafka errors, preserve dlq headers, init coarsetime updater, move offset tracker.
