+++ rust-arroyo/src/processing/stream/mod.rs
@@ -1,6 +1,7 @@
 mod pipeline_envelope;
 mod ext;
 pub mod handlers;
+pub mod offset_tracker;
 pub mod source;
 mod stage;
 
@@ -10,5 +11,6 @@ pub use handlers::{
     DlqHandler, KafkaProducerHandler, LogHandler,
     RejectionHandler, RejectionMetadata, NextHandler,
 };
+pub use offset_tracker::{OffsetCommitter, OffsetTracker};
 pub use source::{KafkaSource, PullSource};
 pub use stage::{RejectionReason, Stage, StageResult};
